requirements.txt
gpt-researcher
fastapi
uvicorn
python-dotenv